+++ src/components/template/products/Facility/iconAndTextMapping.tsx
